feat: enhance text formatting capabilities by adding support for superscript and subscript styles in DOCX processing
